猎人大厅接口

+++ xsl_user_dao/src/main/java/mapper/XslOrderMapper.java
@@ -1,13 +1,13 @@
 package mapper;
 
-import example.XslOrderExample;
 import org.apache.ibatis.annotations.Param;
 import pojo.XslOrder;
+import example.XslOrderExample;
 
 import java.util.List;
 
 public interface XslOrderMapper {
-    int countByExample(XslOrderExample example);
+    long countByExample(XslOrderExample example);
 
     int deleteByExample(XslOrderExample example);
